+++ CHANGELOG.md
@@ -2,6 +2,8 @@
   * **added** serialport plugin, works in electron and Chrome browser
   * **fixed** schematics and enbled it for browser
   * **fixed** display micropython options after port selection in browser
+  * **fixed** schematics image layout
+  * **fixed** notebook images display
 ## Wyliodrin STUDIO 2.2.1-beta
   * **new** Raspberry Pi images (with docker)
   * **improved** docker deployment